email_processor/requirements.txt
ibm-watsonx-orchestrate
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client

